improve keyboard shortcut feature handling with a dedicated hook + add doc
